% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias.R
\name{Bernoulli}
\alias{Bernoulli}
\title{Tensorflow distribution Bernoulli}
\usage{
Bernoulli(...)
}
\description{
Alias for tensorflow distribution Bernoulli. See the classes section of \url{https://www.tensorflow.org/api_docs/python/tf/contrib/distributions}.
}
\seealso{
\url{https://www.tensorflow.org/api_docs/python/tf/contrib/distributions}
}
