% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias.R
\name{BernoulliWithSigmoidP}
\alias{BernoulliWithSigmoidP}
\title{Tensorflow distribution BernoulliWithSigmoidP}
\usage{
BernoulliWithSigmoidP(...)
}
\description{
Alias for tensorflow distribution BernoulliWithSigmoidP. See the classes section of \url{https://www.tensorflow.org/api_docs/python/tf/contrib/distributions}.
}
\seealso{
\url{https://www.tensorflow.org/api_docs/python/tf/contrib/distributions}
}
