% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias.R
\name{Beta}
\alias{Beta}
\title{Tensorflow distribution Beta}
\usage{
Beta(...)
}
\description{
Alias for tensorflow distribution Beta. See the classes section of \url{https://www.tensorflow.org/api_docs/python/tf/contrib/distributions}.
}
\seealso{
\url{https://www.tensorflow.org/api_docs/python/tf/contrib/distributions}
}
