% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias.R
\name{BetaWithSoftplusAB}
\alias{BetaWithSoftplusAB}
\title{Tensorflow distribution BetaWithSoftplusAB}
\usage{
BetaWithSoftplusAB(...)
}
\description{
Alias for tensorflow distribution BetaWithSoftplusAB. See the classes section of \url{https://www.tensorflow.org/api_docs/python/tf/contrib/distributions}.
}
\seealso{
\url{https://www.tensorflow.org/api_docs/python/tf/contrib/distributions}
}
