% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias.R
\name{Binomial}
\alias{Binomial}
\title{Tensorflow distribution Binomial}
\usage{
Binomial(...)
}
\description{
Alias for tensorflow distribution Binomial. See the classes section of \url{https://www.tensorflow.org/api_docs/python/tf/contrib/distributions}.
}
\seealso{
\url{https://www.tensorflow.org/api_docs/python/tf/contrib/distributions}
}
