% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias.R
\name{Chi2}
\alias{Chi2}
\title{Tensorflow distribution Chi2}
\usage{
Chi2(...)
}
\description{
Alias for tensorflow distribution Chi2. See the classes section of \url{https://www.tensorflow.org/api_docs/python/tf/contrib/distributions}.
}
\seealso{
\url{https://www.tensorflow.org/api_docs/python/tf/contrib/distributions}
}
