% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias.R
\name{Dirichlet}
\alias{Dirichlet}
\title{Tensorflow distribution Dirichlet}
\usage{
Dirichlet(...)
}
\description{
Alias for tensorflow distribution Dirichlet. See the classes section of \url{https://www.tensorflow.org/api_docs/python/tf/contrib/distributions}.
}
\seealso{
\url{https://www.tensorflow.org/api_docs/python/tf/contrib/distributions}
}
