% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias.R
\name{DirichletMultinomial}
\alias{DirichletMultinomial}
\title{Tensorflow distribution DirichletMultinomial}
\usage{
DirichletMultinomial(...)
}
\description{
Alias for tensorflow distribution DirichletMultinomial. See the classes section of \url{https://www.tensorflow.org/api_docs/python/tf/contrib/distributions}.
}
\seealso{
\url{https://www.tensorflow.org/api_docs/python/tf/contrib/distributions}
}
