% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias.R
\name{Distribution}
\alias{Distribution}
\title{Tensorflow distribution Distribution}
\usage{
Distribution(...)
}
\description{
Alias for tensorflow distribution Distribution. See the classes section of \url{https://www.tensorflow.org/api_docs/python/tf/contrib/distributions}.
}
\seealso{
\url{https://www.tensorflow.org/api_docs/python/tf/contrib/distributions}
}
