% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias.R
\name{ExponentialWithSoftplusLam}
\alias{ExponentialWithSoftplusLam}
\title{Tensorflow distribution ExponentialWithSoftplusLam}
\usage{
ExponentialWithSoftplusLam(...)
}
\description{
Alias for tensorflow distribution ExponentialWithSoftplusLam. See the classes section of \url{https://www.tensorflow.org/api_docs/python/tf/contrib/distributions}.
}
\seealso{
\url{https://www.tensorflow.org/api_docs/python/tf/contrib/distributions}
}
