% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias.R
\name{Gamma}
\alias{Gamma}
\title{Tensorflow distribution Gamma}
\usage{
Gamma(...)
}
\description{
Alias for tensorflow distribution Gamma. See the classes section of \url{https://www.tensorflow.org/api_docs/python/tf/contrib/distributions}.
}
\seealso{
\url{https://www.tensorflow.org/api_docs/python/tf/contrib/distributions}
}
