% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias.R
\name{GammaWithSoftplusAlphaBeta}
\alias{GammaWithSoftplusAlphaBeta}
\title{Tensorflow distribution GammaWithSoftplusAlphaBeta}
\usage{
GammaWithSoftplusAlphaBeta(...)
}
\description{
Alias for tensorflow distribution GammaWithSoftplusAlphaBeta. See the classes section of \url{https://www.tensorflow.org/api_docs/python/tf/contrib/distributions}.
}
\seealso{
\url{https://www.tensorflow.org/api_docs/python/tf/contrib/distributions}
}
