% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{installTF}
\alias{installTF}
\title{Install TensorFlow and TensorFlow Probability}
\usage{
installTF()
}
\description{
Install the python packages required by sgmcmc, including TensorFlow and TensorFlow probability.
 Uses the tensorflow::install_tensorflow function.
}
