% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias.R
\name{InverseGammaWithSoftplusAlphaBeta}
\alias{InverseGammaWithSoftplusAlphaBeta}
\title{Tensorflow distribution InverseGammaWithSoftplusAlphaBeta}
\usage{
InverseGammaWithSoftplusAlphaBeta(...)
}
\description{
Alias for tensorflow distribution InverseGammaWithSoftplusAlphaBeta. See the classes section of \url{https://www.tensorflow.org/api_docs/python/tf/contrib/distributions}.
}
\seealso{
\url{https://www.tensorflow.org/api_docs/python/tf/contrib/distributions}
}
