% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias.R
\name{Laplace}
\alias{Laplace}
\title{Tensorflow distribution Laplace}
\usage{
Laplace(...)
}
\description{
Alias for tensorflow distribution Laplace. See the classes section of \url{https://www.tensorflow.org/api_docs/python/tf/contrib/distributions}.
}
\seealso{
\url{https://www.tensorflow.org/api_docs/python/tf/contrib/distributions}
}
