% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias.R
\name{LaplaceWithSoftplusScale}
\alias{LaplaceWithSoftplusScale}
\title{Tensorflow distribution LaplaceWithSoftplusScale}
\usage{
LaplaceWithSoftplusScale(...)
}
\description{
Alias for tensorflow distribution LaplaceWithSoftplusScale. See the classes section of \url{https://www.tensorflow.org/api_docs/python/tf/contrib/distributions}.
}
\seealso{
\url{https://www.tensorflow.org/api_docs/python/tf/contrib/distributions}
}
