% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.r
\name{mcmcStep}
\alias{mcmcStep}
\title{Perform a single MCMC step using the dynamics defined in the sgmcmc object}
\usage{
mcmcStep(sgmcmc, sess)
}
\arguments{
\item{sgmcmc}{or sgmcmcCV object as returned by sgld, sghmc or sgnht functions}

\item{sess}{a Tensorflow Session}
}
\description{
If you want to perform Gibbs updates alongside SGMCMC, you'll need to be able to declare
 your own for loop, run those Gibbs updates for each iteration and then run a sinlge MCMC step.
 This function performs a single update of the sgmcmc dynamics declared in the sgmcmc object,
 and updates the sgmcmc$params objects accordingly.
}
\examples{
Tutorials available at [link to be added]

}
