% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias.R
\name{MultivariateNormalCholesky}
\alias{MultivariateNormalCholesky}
\title{Tensorflow distribution MultivariateNormalCholesky}
\usage{
MultivariateNormalCholesky(...)
}
\description{
Alias for tensorflow distribution MultivariateNormalCholesky. See the classes section of \url{https://www.tensorflow.org/api_docs/python/tf/contrib/distributions}.
}
\seealso{
\url{https://www.tensorflow.org/api_docs/python/tf/contrib/distributions}
}
