% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias.R
\name{MultivariateNormalDiag}
\alias{MultivariateNormalDiag}
\title{Tensorflow distribution MultivariateNormalDiag}
\usage{
MultivariateNormalDiag(...)
}
\description{
Alias for tensorflow distribution MultivariateNormalDiag. See the classes section of \url{https://www.tensorflow.org/api_docs/python/tf/contrib/distributions}.
}
\seealso{
\url{https://www.tensorflow.org/api_docs/python/tf/contrib/distributions}
}
