% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias.R
\name{MultivariateNormalDiagPlusVDVT}
\alias{MultivariateNormalDiagPlusVDVT}
\title{Tensorflow distribution MultivariateNormalDiagPlusVDVT}
\usage{
MultivariateNormalDiagPlusVDVT(...)
}
\description{
Alias for tensorflow distribution MultivariateNormalDiagPlusVDVT. See the classes section of \url{https://www.tensorflow.org/api_docs/python/tf/contrib/distributions}.
}
\seealso{
\url{https://www.tensorflow.org/api_docs/python/tf/contrib/distributions}
}
