% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias.R
\name{MultivariateNormalDiagWithSoftplusStDev}
\alias{MultivariateNormalDiagWithSoftplusStDev}
\title{Tensorflow distribution MultivariateNormalDiagWithSoftplusStDev}
\usage{
MultivariateNormalDiagWithSoftplusStDev(...)
}
\description{
Alias for tensorflow distribution MultivariateNormalDiagWithSoftplusStDev. See the classes section of \url{https://www.tensorflow.org/api_docs/python/tf/contrib/distributions}.
}
\seealso{
\url{https://www.tensorflow.org/api_docs/python/tf/contrib/distributions}
}
