% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias.R
\name{NormalWithSoftplusSigma}
\alias{NormalWithSoftplusSigma}
\title{Tensorflow distribution NormalWithSoftplusSigma}
\usage{
NormalWithSoftplusSigma(...)
}
\description{
Alias for tensorflow distribution NormalWithSoftplusSigma. See the classes section of \url{https://www.tensorflow.org/api_docs/python/tf/contrib/distributions}.
}
\seealso{
\url{https://www.tensorflow.org/api_docs/python/tf/contrib/distributions}
}
