% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controlVariates.r
\name{optUpdate}
\alias{optUpdate}
\title{Perform one optimization step on control variate parameters}
\usage{
optUpdate(sess, sgmcmc)
}
\arguments{
\item{sess}{a Tensorflow Session}

\item{sgmcmcCV}{object as returned by sgldCV, sghmcCV or sgnhtCV functions}
}
\description{
If you want to perform Gibbs updates alongside SGMCMCCV, you'll need to be able to declare
 your own for loop, run those Gibbs updates for that iteration and then run a single MCMC
 or optimization step alongside those updates.
 This function performs a single optimization step on the control variate parameters.
 This initial optimization procedure is needed before SGMCMCCV is applied in order to
 ensure the control variate parameters estimate the mode well.
}
\examples{
Tutorials available at [link to be added]

}
