% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias.R
\name{Poisson}
\alias{Poisson}
\title{Tensorflow distribution Poisson}
\usage{
Poisson(...)
}
\description{
Alias for tensorflow distribution Poisson. See the classes section of \url{https://www.tensorflow.org/api_docs/python/tf/contrib/distributions}.
}
\seealso{
\url{https://www.tensorflow.org/api_docs/python/tf/contrib/distributions}
}
