% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias.R
\name{QuantizedDistribution}
\alias{QuantizedDistribution}
\title{Tensorflow distribution QuantizedDistribution}
\usage{
QuantizedDistribution(...)
}
\description{
Alias for tensorflow distribution QuantizedDistribution. See the classes section of \url{https://www.tensorflow.org/api_docs/python/tf/contrib/distributions}.
}
\seealso{
\url{https://www.tensorflow.org/api_docs/python/tf/contrib/distributions}
}
