% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias.R
\name{RegisterKL}
\alias{RegisterKL}
\title{Tensorflow distribution RegisterKL}
\usage{
RegisterKL(...)
}
\description{
Alias for tensorflow distribution RegisterKL. See the classes section of \url{https://www.tensorflow.org/api_docs/python/tf/contrib/distributions}.
}
\seealso{
\url{https://www.tensorflow.org/api_docs/python/tf/contrib/distributions}
}
