% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sghmc.r
\name{runSGHMC}
\alias{runSGHMC}
\title{Stochastic Gradient Hamiltonian Monte Carlo}
\usage{
runSGHMC(logLik, logPrior, data, params, eta, alpha, L, n, nIters = 10^4,
  verbose = TRUE)
}
\arguments{
\item{logLik}{function which takes parameters and data 
(list of tensorflow variables and placeholders respectively) as input. 
It should return a tensorflow expression which defines the log likelihood of the model.}

\item{logPrior}{function which takes parameters and data 
(list of tensorflow variables and placeholders respectively) as input. 
The function should return a tensorflow tensor which defines the log prior of the model.}

\item{data}{list of R arrays which defines the datasets for the problem.
The names in the list should correspond to those referred to in the logLik and logPrior functions}

\item{params}{list of R arrays which define the starting point of each parameter.
The names in the list should correspond to those referred to in the logLik and logPrior functions}

\item{eta}{list of numeric values corresponding to the SGHMC eta stepsize terms,
as defined in the paper given in the references. One value should be given 
for each parameter in params, the names should correspond to those in params.}

\item{alpha}{list of numeric values corresponding to the SGHMC alpha momentum terms,
as defined in the paper given in the references. One value should be given 
for each parameter in params, the names should correspond to those in params.}

\item{L}{integer specifying the trajectory of the simulation, as defined in the main reference.}

\item{n}{minibatch size, assumed integer.}

\item{nIters}{number of iterations of SGLD to perform, optional, assumed integer, default 10^4.}

\item{verbose}{whether to print algorithm progress or not, assumed BOOLEAN, default TRUE.}
}
\value{
List of arrays for each parameter containing the MCMC chain.
 Dimension of the form (nIters,paramDim1,paramDim2,...)
}
\description{
Simulates from the posterior defined by the functions logLik & logPrior using
 stochastic gradient Hamiltonian Monte Carlo. The function uses TensorFlow, so needs
 Tensorflow for python installed.
}
\details{
@references Chen, T., Fox, E. B., & Guestrin, C. (2014). 
     stochastic gradient Hamiltonian Monte Carlo. In ICML (pp. 1683-1691).
}
\examples{
Tutorials available at [link to be added]

}
