% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sghmc.r
\name{runSGHMCCV}
\alias{runSGHMCCV}
\title{Stochastic Gradient Hamiltonian Monte Carlo with Control Variates}
\usage{
runSGHMCCV(logLik, logPrior, data, params, eta, alpha, L, optStepsize, n,
  nIters = 10^4, nItersOpt = 10^4, verbose = TRUE)
}
\arguments{
\item{logLik}{function which takes parameters and data 
(list of tensorflow variables and placeholders respectively) as input. 
It should return a tensorflow expression which defines the log likelihood of the model.}

\item{logPrior}{function which takes parameters and data 
(list of tensorflow variables and placeholders respectively) as input. 
The function should return a tensorflow tensor which defines the log prior of the model.}

\item{data}{list of R arrays which defines the datasets for the problem.
The names in the list should correspond to those referred to in the logLik and logPrior functions}

\item{params}{list of R arrays which define the starting point of each parameter.
The names in the list should correspond to those referred to in the logLik and logPrior functions}

\item{eta}{list of numeric values corresponding to the SGHMC eta stepsize terms,
as defined in the paper given in the references. One value should be given 
for each parameter in params, the names should correspond to those in params.}

\item{alpha}{list of numeric values corresponding to the SGHMC alpha momentum terms,
as defined in the paper given in the references. One value should be given 
for each parameter in params, the names should correspond to those in params.}

\item{L}{integer specifying the trajectory of the simulation, as defined in the main reference.}

\item{optStepsize}{numeric, stepsize for optimization step to find MAP estimates of parameters.}

\item{n}{minibatch size, assumed integer.}

\item{nIters}{integer, number of iterations of SGLD to perform, optional, default 10^4.}

\item{nItersOpt}{integer, number of iterations of initial optimization to perform, 
optional, default 10^4.}

\item{verbose}{Boolean, whether to print algorithm progress or not, default TRUE.}
}
\value{
List of arrays for each parameter containing the MCMC chain.
 Dimension of the form (nIters,paramDim1,paramDim2,...)
}
\description{
Simulates from the posterior defined by the functions logLik & logPrior using 
 stochastic gradient Hamiltonian Monte Carlo with an improved gradient estimate 
 that is calculated using control variates.
}
\details{
@references Baker, J., Fearnhead, P., Fox, E. B., & Nemeth, C. (2017) 
     control variates for stochastic gradient Langevin dynamics. Preprint.
 @references Chen, T., Fox, E. B., & Guestrin, C. (2014). 
     stochastic gradient Hamiltonian Monte Carlo. In ICML (pp. 1683-1691).
}
\examples{
Tutorials available at [link to be added]

}
