% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgnht.r
\name{runSGNHT}
\alias{runSGNHT}
\title{Stochastic Gradient Nose Hoover Thermostat}
\usage{
runSGNHT(logLik, logPrior, data, params, eta, a, n, nIters = 10^4,
  verbose = TRUE)
}
\arguments{
\item{logLik}{function which takes parameters and data 
(list of tensorflow variables and placeholders respectively) as input. 
It should return a tensorflow expression which defines the log likelihood of the model.}

\item{logPrior}{function which takes parameters and data 
(list of tensorflow variables and placeholders respectively) as input. 
The function should return a tensorflow tensor which defines the log prior of the model.}

\item{data}{list of R arrays which defines the datasets for the problem.
The names in the list should correspond to those referred to in the logLik and logPrior functions}

\item{params}{list of R arrays which define the starting point of each parameter.
The names in the list should correspond to those referred to in the logLik and logPrior functions}

\item{eta}{list of numeric values corresponding to the SGNHT eta stepsize terms,
as defined in the paper given in the references, Appendix F. One value should be given 
for each parameter in params, the names should correspond to those in params.}

\item{a}{list of numeric values corresponding to the SGNHT a diffusion factors,
as defined in the paper given in the references, Appendix F. One value should be given 
for each parameter in params, the names should correspond to those in params.}

\item{n}{minibatch size, assumed integer.}

\item{nIters}{number of iterations of SGLD to perform, optional, assumed integer, default 10^4.}

\item{verbose}{whether to print algorithm progress or not, assumed BOOLEAN, default TRUE.}
}
\value{
List of arrays for each parameter containing the MCMC chain.
 Dimension of the form (nIters,paramDim1,paramDim2,...)
}
\description{
Simulates from the posterior defined by the functions logLik & logPrior using
 stochastic gradient Nose Hoover Thermostat.
 When used for parameters that are higher dimensional then vectors, the thermostat
 step is approximated by a tensor contraction.
}
\details{
@references Ding, N., Fang, Y., Babbush, R., Chen, C., Skeel, R. D., & Neven, H. (2014). 
     Bayesian sampling using stochastic gradient thermostats. NIPS (pp. 3203-3211).
}
\examples{
Tutorials available at [link to be added]

}
