% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sghmc.r
\name{sghmc}
\alias{sghmc}
\title{Create Stochastic Gradient Hamiltonian Monte Carlo Object}
\usage{
sghmc(logLik, logPrior, data, params, eta, alpha, L, n, gibbsParams)
}
\arguments{
\item{logLik}{function which takes parameters and data 
(list of tensorflow variables and placeholders respectively) as input. 
It should return a tensorflow expression which defines the log likelihood of the model.}

\item{logPrior}{function which takes parameters and data 
(list of tensorflow variables and placeholders respectively) as input. 
The function should return a tensorflow tensor which defines the log prior of the model.}

\item{data}{list of R numeric arrays which defines the datasets for the problem.
The names in the list should correspond to those referred to in the logLik and logPrior functions}

\item{params}{list of numeric R arrays which define the starting point of each parameter.
The names in the list should correspond to those referred to in the logLik and logPrior functions}

\item{eta}{list of numeric values corresponding to the SGHMC eta stepsize terms,
as defined in the paper given in the references. One value should be given 
for each parameter in params, the names should correspond to those in params.}

\item{alpha}{list of numeric values corresponding to the SGHMC alpha momentum terms,
as defined in the paper given in the references. One value should be given 
for each parameter in params, the names should correspond to those in params.}

\item{L}{integer specifying the trajectory of the simulation, as defined in the main reference.}

\item{n}{integer corresponding to minibatch size}

\item{gibbsParams}{list of tensorflow variables, corresponding to parameters that will be updated
manually using Gibbs steps}
}
\value{
List of arrays for each parameter containing the MCMC chain.
 Dimension of the form (nIters,paramDim1,paramDim2,...)
}
\description{
Creates a stochastic gradient Langevin Dynamics (SGHMC) object which can be passed to mcmcStep
 to simulate from 1 step of SGHMC for the posterior defined by logLik and logPrior.
 This is for advanced use of the library when additional Gibbs steps are needed.
}
\details{
@references Chen, T., Fox, E. B., & Guestrin, C. (2014). 
     stochastic gradient Hamiltonian Monte Carlo. In ICML (pp. 1683-1691).
}
\examples{
Tutorials available at [link to be added]

}
