% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sghmc.r
\name{sghmcCV}
\alias{sghmcCV}
\title{Create Stochastic Gradient Hamiltonian Monte Carlo with Control Variates Object}
\usage{
sghmcCV(logLik, logPrior, data, params, eta, alpha, L, optStepsize, n,
  gibbsParams)
}
\arguments{
\item{logLik}{function which takes parameters and data 
(list of tensorflow variables and placeholders respectively) as input. 
It should return a tensorflow expression which defines the log likelihood of the model.}

\item{logPrior}{function which takes parameters and data 
(list of tensorflow variables and placeholders respectively) as input. 
The function should return a tensorflow tensor which defines the log prior of the model.}

\item{data}{list of R numeric arrays which defines the datasets for the problem.
The names in the list should correspond to those referred to in the logLik and logPrior functions}

\item{params}{list of numeric R arrays which define the starting point of each parameter.
The names in the list should correspond to those referred to in the logLik and logPrior functions}

\item{optStepsize}{numeric, stepsize for optimization step to find MAP estimates of parameters.}

\item{n}{integer corresponding to minibatch size}

\item{gibbsParams}{list of tensorflow variables, corresponding to parameters that will be updated
manually using Gibbs steps}

\item{stepsize}{list of numeric stepsizes corresponding to the SGLD stepsizes for each parameter
The names in the list should correspond to those in params.}
}
\value{
List of arrays for each parameter containing the MCMC chain.
 Dimension of the form (nIters,paramDim1,paramDim2,...)
}
\description{
Creates a stochastic gradient Hamiltonian Monte Carlo with Control Variates (SGHMCCV) object 
 which can be passed to optUpdate and mcmcStep functions to simulate from SGHMCCV
 for the posterior defined by logLik and logPrior.
 This is for advanced use of the library when additional Gibbs steps are needed.
}
\details{
@references Baker, J., Fearnhead, P., Fox, E. B., & Nemeth, C. (2017) 
     control variates for stochastic gradient Langevin dynamics. Preprint.
 @references Chen, T., Fox, E. B., & Guestrin, C. (2014). 
     stochastic gradient Hamiltonian Monte Carlo. In ICML (pp. 1683-1691).
}
\examples{
Tutorials available at [link to be added]

}
