% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias.R
\name{StudentTWithAbsDfSoftplusSigma}
\alias{StudentTWithAbsDfSoftplusSigma}
\title{Tensorflow distribution StudentTWithAbsDfSoftplusSigma}
\usage{
StudentTWithAbsDfSoftplusSigma(...)
}
\description{
Alias for tensorflow distribution StudentTWithAbsDfSoftplusSigma. See the classes section of \url{https://www.tensorflow.org/api_docs/python/tf/contrib/distributions}.
}
\seealso{
\url{https://www.tensorflow.org/api_docs/python/tf/contrib/distributions}
}
