% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias.R
\name{Uniform}
\alias{Uniform}
\title{Tensorflow distribution Uniform}
\usage{
Uniform(...)
}
\description{
Alias for tensorflow distribution Uniform. See the classes section of \url{https://www.tensorflow.org/api_docs/python/tf/contrib/distributions}.
}
\seealso{
\url{https://www.tensorflow.org/api_docs/python/tf/contrib/distributions}
}
