% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias.R
\name{WishartCholesky}
\alias{WishartCholesky}
\title{Tensorflow distribution WishartCholesky}
\usage{
WishartCholesky(...)
}
\description{
Alias for tensorflow distribution WishartCholesky. See the classes section of \url{https://www.tensorflow.org/api_docs/python/tf/contrib/distributions}.
}
\seealso{
\url{https://www.tensorflow.org/api_docs/python/tf/contrib/distributions}
}
