% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgld.r
\name{sgld}
\alias{sgld}
\title{Create Stochastic Gradient Langevin Dynamics Object}
\usage{
sgld(logLik, logPrior, data, params, stepsize, n, gibbsParams)
}
\arguments{
\item{logLik}{function which takes parameters and data 
(list of tensorflow variables and placeholders respectively) as input. 
It should return a tensorflow expression which defines the log likelihood of the model.}

\item{logPrior}{function which takes parameters and data 
(list of tensorflow variables and placeholders respectively) as input. 
The function should return a tensorflow tensor which defines the log prior of the model.}

\item{data}{list of R numeric arrays which defines the datasets for the problem.
The names in the list should correspond to those referred to in the logLik and logPrior functions}

\item{params}{list of numeric R arrays which define the starting point of each parameter.
The names in the list should correspond to those referred to in the logLik and logPrior functions}

\item{stepsize}{list of numeric stepsizes corresponding to the SGLD stepsizes for each parameter
The names in the list should correspond to those in params.}

\item{n}{integer corresponding to minibatch size}

\item{gibbsParams}{list of tensorflow variables, corresponding to parameters that will be updated
manually using Gibbs steps}
}
\value{
List of arrays for each parameter containing the MCMC chain.
 Dimension of the form (nIters,paramDim1,paramDim2,...)
}
\description{
Creates a stochastic gradient Langevin Dynamics (SGLD) object which can be passed to mcmcStep
 to simulate from 1 step of SGLD for the posterior defined by logLik and logPrior.
 This is for advanced use of the library when additional Gibbs steps are needed.
}
\details{
@references Welling, M., & Teh, Y. W. (2011). 
     Bayesian learning via stochastic gradient Langevin dynamics. ICML (pp. 681-688).
}
\examples{
Tutorials available at [link to be added]

}
