% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgnht.R
\name{sgnhtcv}
\alias{sgnhtcv}
\title{Stochastic Gradient Nose Hoover Thermostat with Control Variates}
\usage{
sgnhtcv(logLik, logPrior, data, params, eta, a, optStepsize, n, nIters = 10^4,
  nItersOpt = 10^4, verbose = TRUE)
}
\arguments{
\item{logLik}{function which takes parameters and data 
(list of tensorflow variables and placeholders respectively) as input. 
It should return a tensorflow expression which defines the log likelihood of the model.}

\item{logPrior}{function which takes parameters and data 
(list of tensorflow variables and placeholders respectively) as input. 
The function should return a tensorflow tensor which defines the log prior of the model.}

\item{data}{list of R arrays which defines the datasets for the problem.
The names in the list should correspond to those referred to in the logLik and logPrior functions}

\item{params}{list of R arrays which define the starting point of each parameter.
The names in the list should correspond to those referred to in the logLik and logPrior functions}

\item{eta}{list of numeric values corresponding to the SGNHT eta stepsize terms,
as defined in the paper given in the references, Appendix F. One value should be given 
for each parameter in params, the names should correspond to those in params.}

\item{a}{list of numeric values corresponding to the SGNHT a diffusion factors,
as defined in the paper given in the references, Appendix F. One value should be given 
for each parameter in params, the names should correspond to those in params.}

\item{optStepsize}{numeric, stepsize for optimization step to find MAP estimates of parameters.}

\item{n}{minibatch size, assumed integer.}

\item{nIters}{integer, number of iterations of SGLD to perform, optional, default 10^4.}

\item{nItersOpt}{integer, number of iterations of initial optimization to perform, 
optional, default 10^4.}

\item{verbose}{Boolean, whether to print algorithm progress or not, default TRUE.}
}
\value{
List of arrays for each parameter containing the MCMC chain.
 Dimension of the form (nIters,paramDim1,paramDim2,...)
}
\description{
Simulates from the posterior defined by the functions logLik & logPrior using
 stochastic gradient Nose Hoover Thermostat with an improved gradient estimate 
 that is calculated using control variates.
 When used for parameters that are higher dimensional then vectors, the thermostat
 step is approximated by a tensor contraction.
}
\references{
Baker, J., Fearnhead, P., Fox, E. B., & Nemeth, C. (2017) 
 control variates for stochastic gradient Langevin dynamics. Preprint.

Ding, N., Fang, Y., Babbush, R., Chen, C., Skeel, R. D., & Neven, H. (2014). 
 Bayesian sampling using stochastic gradient thermostats. NIPS (pp. 3203-3211).
}
