% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgnht.r
\name{sgnhtCV}
\alias{sgnhtCV}
\title{Create Stochastic Gradient Nose Hoover Thermostat with Control Variates Object}
\usage{
sgnhtCV(logLik, logPrior, data, params, eta, a, optStepsize, n, gibbsParams,
  n_iters = 10^4)
}
\arguments{
\item{logLik}{function which takes parameters and data 
(list of tensorflow variables and placeholders respectively) as input. 
It should return a tensorflow expression which defines the log likelihood of the model.}

\item{logPrior}{function which takes parameters and data 
(list of tensorflow variables and placeholders respectively) as input. 
The function should return a tensorflow tensor which defines the log prior of the model.}

\item{data}{list of R arrays which defines the datasets for the problem.
The names in the list should correspond to those referred to in the logLik and logPrior functions}

\item{params}{list of R arrays which define the starting point of each parameter.
The names in the list should correspond to those referred to in the logLik and logPrior functions}

\item{eta}{list of numeric values corresponding to the SGNHT eta stepsize terms,
as defined in the paper given in the references, Appendix F. One value should be given 
for each parameter in params, the names should correspond to those in params.}

\item{a}{list of numeric values corresponding to the SGNHT a diffusion factors,
as defined in the paper given in the references, Appendix F. One value should be given 
for each parameter in params, the names should correspond to those in params.}

\item{optStepsize}{numeric, stepsize for optimization step to find MAP estimates of parameters.}

\item{n}{minibatch size, assumed integer.}

\item{gibbsParams}{list of tensorflow variables, corresponding to parameters that will be updated
manually using Gibbs steps}
}
\value{
List of arrays for each parameter containing the MCMC chain.
 Dimension of the form (nIters,paramDim1,paramDim2,...)
}
\description{
Creates a stochastic gradient Nose Hoover Thermostat with Control Variates (SGHMCCV) object 
 which can be passed to optUpdate and mcmcStep functions to simulate from SGNHTCV
 for the posterior defined by logLik and logPrior.
 This is for advanced use of the library when additional Gibbs steps are needed.
}
\examples{
Tutorials available at [link to be added]

}
\references{
Baker, J., Fearnhead, P., Fox, E. B., & Nemeth, C. (2017) 
 control variates for stochastic gradient Langevin dynamics. Preprint.

Ding, N., Fang, Y., Babbush, R., Chen, C., Skeel, R. D., & Neven, H. (2014). 
 Bayesian sampling using stochastic gradient thermostats. NIPS (pp. 3203-3211).
}
